fix: move ServiceClientSampleCodeComposer to composer.samplecode
