+++ src/main/java/com/google/api/generator/gapic/composer/samplecode/BUILD.bazel
@@ -17,6 +17,7 @@ java_library(
         "//src/main/java/com/google/api/generator/engine/writer",
         "//src/main/java/com/google/api/generator/gapic/model",
         "//src/main/java/com/google/api/generator/gapic/utils",
+        "@com_google_api_gax_java//gax",
         "@google_java_format_all_deps//jar",
     ],
 )
